This PR introduces pluggable scan items in a similar way to dialogs and product features.

@bdunne Please review

cc @agrare @pemcg @jrafanie