@Fryguy 2019-10-10
Technically this improves seeding performance as well, but barely any :)

| | first seed | second seed |
| --- | --- | --- |
| before | 161.8 ms | 111.4 ms |
| after | 104.8 ms | 24.0 ms |
| savings | 35.2% | 78.5% |