@simaishi 2019-11-18
Ivanchuk backport details:

```
$ git log -1
commit 75d79eadf0c5a291e926e6db5557d2d6bccaf5e8
Author: Brandon Dunne <bdunne@redhat.com>
Date:   Mon Oct 14 14:15:39 2019 -0400

    Merge pull request #19388 from Fryguy/pluggable_scan_items
    
    Pluggable ScanItems
    
    (cherry picked from commit c0664f72920d8c50019144288e8dff11466bab23)
    
    Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1773703
```